Do not load all weather icons on startup
The weather widget loads an enormous number of icons. It takes
more than half a minute on my EeePC 901 to do it. To speed up
the startup process it is reasonable to delay loading icons
until they are needed.

This patch introduces find_icon_file() function which checks a table
of cached icon paths, returns an entry if it is present or, seeks
for the requested icon and creates a new cache entry.

Signed-off-by: Łukasz Stelmach <stlman@poczta.fm>
Signed-off-by: Tuomas Jormola <tj@solitudo.net>